% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreDB.R
\name{asCompadreDB}
\alias{asCompadreDB}
\title{Convert legacy COM(P)ADRE database object to CompadreDB}
\usage{
asCompadreDB(from)
}
\arguments{
\item{from}{A legacy COM(P)ADRE database}
}
\value{
A CompadreDB object
}
\description{
Convert a legacy COM(P)ADRE database object (of class 'list') to a CompadreDB
object.
}
\examples{
Compadre <- asCompadreDB(CompadreLegacy)

}
\author{
Iain M. Stott
}
