% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_fetch.R
\name{cdb_fetch}
\alias{cdb_fetch}
\title{Fetch the COM(P)ADRE database from compadre-db.org or a local file}
\usage{
cdb_fetch(cdb)
}
\arguments{
\item{cdb}{Either "comadre" or "compadre" (case insensitive) to fetch the
most recent database from \url{https://compadre-db.org}, or a path to an
existing COMPADRE database (i.e. .RData file) stored on the local machine.}
}
\value{
A CompadreDB object
}
\description{
Fetches the current version of a COM(P)ADRE database from
\url{https://compadre-db.org}, or load any version stored in a local .RData
file.
}
\examples{
\dontrun{
compadre <- cdb_fetch('compadre')
}
}
\seealso{
Other {data acquisition}: 
\code{\link{cdb_metadata}()}
}
\author{
Patrick M. Barks <patrick.barks@gmail.com>
}
\concept{{data acquisition}}
