% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_flatten.R
\name{cdb_flatten}
\alias{cdb_flatten}
\title{Convert a COM(P)ADRE database object to a flat data frame}
\usage{
cdb_flatten(cdb, onlyMatA = FALSE)
}
\arguments{
\item{cdb}{A CompadreDB object}

\item{onlyMatA}{A logical value (TRUE/FALSE) indicating whether ONLY the full
projection matrix \code{matA} should be included in the flattened data
frame}
}
\value{
The \code{data.frame} from the data slot of \code{cdb}, but with
  additional columns appended for the matrix stage information and the
  matrices themselves, both in string format.
}
\description{
Converts a CompadreDB object to a flat data frame, by converting each matrix
and associated matrixClass information to a string.
}
\examples{
CompadreFlat <- cdb_flatten(Compadre, onlyMatA = FALSE)

}
\seealso{
string_to_mat
}
\author{
Owen R. Jones <jones@biology.sdu.dk>
}
