% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDBs.R
\name{cdb_merge}
\alias{cdb_merge}
\title{Merge two COM(P)ADRE databases together}
\usage{
cdb_merge(cdb1, cdb2)
}
\arguments{
\item{cdb1, cdb2}{CompadreDB objects}
}
\value{
A CompadreDB object containing both databases
}
\description{
Merges two CompadreDB objects via a row-bind of the data slots.
}
\examples{
Compadre1 <- subset(Compadre, Continent == "Asia")
Compadre2 <- subset(Compadre, Continent == "Africa")

cdb_merge(Compadre1, Compadre2)

}
\author{
Sam Levin
}
