% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_subset_wild.R
\name{cdb_subset_wild}
\alias{cdb_subset_wild}
\title{Subset a COM(P)ADRE database to wild, unmanipulated populations}
\usage{
cdb_subset_wild(cdb)
}
\arguments{
\item{cdb}{A CompadreDB object}
}
\value{
A CompadreDB object reflecting the subset of \code{cdb} corresponding
  to wild (\code{MatrixCaptivity = "W"}), unmanipulated
  (\code{MatrixTreatment == "Unmanipulated"}) populations.
}
\description{
Subset a COM(P)ADRE database to the rows reflecting wild, unmanipulated
populations \code{(MatrixCaptivity == "W" & MatrixTreatment ==
"Unmanipulated")}.

Note that such a subset may be overly stringent, as it excludes populations
subject to "treatments" (i.e. \code{MatrixTreatment != "Unmanipulated"}) that
may not necessitate exclusion from a given analysis (e.g. "treatment" may
reflect grazing status, time since fire, etc.).
}
\examples{
CompadreWild <- cdb_subset_wild(Compadre)

}
\author{
Patrick M. Barks <patrick.barks@gmail.com>
}
