% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_tidy.R
\name{cdb_tidy}
\alias{cdb_tidy}
\title{Convert a COM(P)ADRE database to a tidy data frame}
\usage{
cdb_tidy(cdb)
}
\arguments{
\item{cdb}{A CompadreDB object}
}
\value{
A tibble-style data frame based on the data slot of \code{cdb}, but
  with the list-column \code{mat} split into 6 separate list-columns (for
  matrices \code{matA}, \code{matU}, \code{matF}, \code{matC}, and vectors
  \code{MatrixClassAuthor}, and \code{MatrixClassOrganized}).
}
\description{
Converts a CompadreDB object to a tidy data frame by extracting the data
slot, and splitting the \code{mat} column into separate list-columns for each
component (matrices \code{matA}, \code{matU}, \code{matF}, \code{matC}, and
vectors \code{MatrixClassAuthor}, and \code{MatrixClassOrganized}).
}
\examples{
CompadreTidy <- cdb_tidy(Compadre)

}
\author{
Patrick M. Barks <patrick.barks@gmail.com>
}
