% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_vec_studies.R
\name{cdb_vec_studies}
\alias{cdb_vec_studies}
\title{Extract a vector of unique study identifiers from a COM(P)ADRE database}
\usage{
cdb_vec_studies(cdb, columns = c("Authors", "Journal", "YearPublication",
  "DOI.ISBN"))
}
\arguments{
\item{cdb}{A CompadreDB object}

\item{columns}{A vector of column names for which unique combinations should
be identified. Defaults to \code{c("Authors", "Journal", "YearPublication",
"DOI.ISBN")}.}
}
\value{
Vector of integer study identifiers corresponding to the rows of
  \code{cdb}.
}
\description{
Extracts a vector of integer study identifiers corresponding to the rows of a
CompadreDB object, based on unique combinations of the columns 'Authors',
'Journal', 'YearPublication', and 'DOI.ISBN' (or optionally, a different set
of columns supplied by the user).
}
\examples{
Compadre$StudyID <- cdb_vec_studies(Compadre)

}
\author{
Patrick Barks <patrick.barks@gmail.com>
}
