% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassUnionMethods.R
\docType{class}
\name{CompadreMorData-class}
\alias{CompadreMorData-class}
\alias{matA}
\alias{matA,CompadreMorData-method}
\alias{matU}
\alias{matU,CompadreMorData-method}
\alias{matF}
\alias{matF,CompadreMorData-method}
\alias{matC}
\alias{matC,CompadreMorData-method}
\alias{matrixClass}
\alias{matrixClass,CompadreMorData-method}
\alias{metadata}
\alias{metadata,CompadreData-method}
\alias{SpeciesAuthor}
\alias{SpeciesAuthor,CompadreData-method}
\alias{SpeciesAccepted}
\alias{SpeciesAccepted,CompadreData-method}
\alias{CommonName}
\alias{CommonName,CompadreData-method}
\alias{Infraspecific}
\alias{Infraspecific,CompadreData-method}
\alias{Genus}
\alias{Genus,CompadreData-method}
\alias{GenusAccepted}
\alias{GenusAccepted,CompadreData-method}
\alias{GenusAuthor}
\alias{GenusAuthor,CompadreData-method}
\alias{Family}
\alias{Family,CompadreData-method}
\alias{Order}
\alias{Order,CompadreData-method}
\alias{Class}
\alias{Class,CompadreData-method}
\alias{Phylum}
\alias{Phylum,CompadreData-method}
\alias{Kingdom}
\alias{Kingdom,CompadreData-method}
\alias{OrganismType}
\alias{OrganismType,CompadreData-method}
\alias{DicotMonoc}
\alias{DicotMonoc,CompadreData-method}
\alias{AngioGymno}
\alias{AngioGymno,CompadreData-method}
\alias{Authors}
\alias{Authors,CompadreData-method}
\alias{Journal}
\alias{Journal,CompadreData-method}
\alias{YearPublication}
\alias{YearPublication,CompadreData-method}
\alias{DOI.ISBN}
\alias{DOI.ISBN,CompadreData-method}
\alias{AdditionalSource}
\alias{AdditionalSource,CompadreData-method}
\alias{StudyDuration}
\alias{StudyDuration,CompadreData-method}
\alias{StudyStart}
\alias{StudyStart,CompadreData-method}
\alias{StudyEnd}
\alias{StudyEnd,CompadreData-method}
\alias{AnnualPeriodicity}
\alias{AnnualPeriodicity,CompadreData-method}
\alias{NumberPopulations}
\alias{NumberPopulations,CompadreData-method}
\alias{MatrixCriteriaSize}
\alias{MatrixCriteriaSize,CompadreData-method}
\alias{MatrixCriteriaOntogeny}
\alias{MatrixCriteriaOntogeny,CompadreData-method}
\alias{MatrixCriteriaAge}
\alias{MatrixCriteriaAge,CompadreData-method}
\alias{MatrixPopulation}
\alias{MatrixPopulation,CompadreData-method}
\alias{Lat}
\alias{Lat,CompadreData-method}
\alias{Lon}
\alias{Lon,CompadreData-method}
\alias{Altitude}
\alias{Altitude,CompadreData-method}
\alias{Country}
\alias{Country,CompadreData-method}
\alias{Continent}
\alias{Continent,CompadreData-method}
\alias{Ecoregion}
\alias{Ecoregion,CompadreData-method}
\alias{StudiedSex}
\alias{StudiedSex,CompadreData-method}
\alias{MatrixComposite}
\alias{MatrixComposite,CompadreData-method}
\alias{MatrixTreatment}
\alias{MatrixTreatment,CompadreData-method}
\alias{MatrixCaptivity}
\alias{MatrixCaptivity,CompadreData-method}
\alias{MatrixStartYear}
\alias{MatrixStartYear,CompadreData-method}
\alias{MatrixStartSeason}
\alias{MatrixStartSeason,CompadreData-method}
\alias{MatrixStartMonth}
\alias{MatrixStartMonth,CompadreData-method}
\alias{MatrixEndYear}
\alias{MatrixEndYear,CompadreData-method}
\alias{MatrixEndSeason}
\alias{MatrixEndSeason,CompadreData-method}
\alias{MatrixEndMonth}
\alias{MatrixEndMonth,CompadreData-method}
\alias{MatrixSplit}
\alias{MatrixSplit,CompadreData-method}
\alias{MatrixFec}
\alias{MatrixFec,CompadreData-method}
\alias{Observation}
\alias{Observation,CompadreData-method}
\alias{MatrixDimension}
\alias{MatrixDimension,CompadreData-method}
\alias{SurvivalIssue}
\alias{SurvivalIssue,CompadreData-method}
\title{Accessor methods for CompadreM and CompadreData objects}
\usage{
matA(object)

\S4method{matA}{CompadreMorData}(object)

matU(object)

\S4method{matU}{CompadreMorData}(object)

matF(object)

\S4method{matF}{CompadreMorData}(object)

matC(object)

\S4method{matC}{CompadreMorData}(object)

matrixClass(object)

\S4method{matrixClass}{CompadreMorData}(object)

metadata(object, ...)

\S4method{metadata}{CompadreData}(object, ...)

SpeciesAuthor(object)

\S4method{SpeciesAuthor}{CompadreData}(object)

SpeciesAccepted(object)

\S4method{SpeciesAccepted}{CompadreData}(object)

CommonName(object)

\S4method{CommonName}{CompadreData}(object)

Infraspecific(object)

\S4method{Infraspecific}{CompadreData}(object)

Genus(object)

\S4method{Genus}{CompadreData}(object)

GenusAccepted(object)

\S4method{GenusAccepted}{CompadreData}(object)

GenusAuthor(object)

\S4method{GenusAuthor}{CompadreData}(object)

Family(object)

\S4method{Family}{CompadreData}(object)

Order(object)

\S4method{Order}{CompadreData}(object)

Class(object)

\S4method{Class}{CompadreData}(object)

Phylum(object)

\S4method{Phylum}{CompadreData}(object)

Kingdom(object)

\S4method{Kingdom}{CompadreData}(object)

OrganismType(object)

\S4method{OrganismType}{CompadreData}(object)

DicotMonoc(object)

\S4method{DicotMonoc}{CompadreData}(object)

AngioGymno(object)

\S4method{AngioGymno}{CompadreData}(object)

Authors(object)

\S4method{Authors}{CompadreData}(object)

Journal(object)

\S4method{Journal}{CompadreData}(object)

YearPublication(object)

\S4method{YearPublication}{CompadreData}(object)

DOI.ISBN(object)

\S4method{DOI.ISBN}{CompadreData}(object)

AdditionalSource(object)

\S4method{AdditionalSource}{CompadreData}(object)

StudyDuration(object)

\S4method{StudyDuration}{CompadreData}(object)

StudyStart(object)

\S4method{StudyStart}{CompadreData}(object)

StudyEnd(object)

\S4method{StudyEnd}{CompadreData}(object)

AnnualPeriodicity(object)

\S4method{AnnualPeriodicity}{CompadreData}(object)

NumberPopulations(object)

\S4method{NumberPopulations}{CompadreData}(object)

MatrixCriteriaSize(object)

\S4method{MatrixCriteriaSize}{CompadreData}(object)

MatrixCriteriaOntogeny(object)

\S4method{MatrixCriteriaOntogeny}{CompadreData}(object)

MatrixCriteriaAge(object)

\S4method{MatrixCriteriaAge}{CompadreData}(object)

MatrixPopulation(object)

\S4method{MatrixPopulation}{CompadreData}(object)

Lat(object)

\S4method{Lat}{CompadreData}(object)

Lon(object)

\S4method{Lon}{CompadreData}(object)

Altitude(object)

\S4method{Altitude}{CompadreData}(object)

Country(object)

\S4method{Country}{CompadreData}(object)

Continent(object)

\S4method{Continent}{CompadreData}(object)

Ecoregion(object)

\S4method{Ecoregion}{CompadreData}(object)

StudiedSex(object)

\S4method{StudiedSex}{CompadreData}(object)

MatrixComposite(object)

\S4method{MatrixComposite}{CompadreData}(object)

MatrixTreatment(object)

\S4method{MatrixTreatment}{CompadreData}(object)

MatrixCaptivity(object)

\S4method{MatrixCaptivity}{CompadreData}(object)

MatrixStartYear(object)

\S4method{MatrixStartYear}{CompadreData}(object)

MatrixStartSeason(object)

\S4method{MatrixStartSeason}{CompadreData}(object)

MatrixStartMonth(object)

\S4method{MatrixStartMonth}{CompadreData}(object)

MatrixEndYear(object)

\S4method{MatrixEndYear}{CompadreData}(object)

MatrixEndSeason(object)

\S4method{MatrixEndSeason}{CompadreData}(object)

MatrixEndMonth(object)

\S4method{MatrixEndMonth}{CompadreData}(object)

MatrixSplit(object)

\S4method{MatrixSplit}{CompadreData}(object)

MatrixFec(object)

\S4method{MatrixFec}{CompadreData}(object)

Observation(object)

\S4method{Observation}{CompadreData}(object)

MatrixDimension(object)

\S4method{MatrixDimension}{CompadreData}(object)

SurvivalIssue(object)

\S4method{SurvivalIssue}{CompadreData}(object)
}
\description{
This page describes methods for both the CompadreM and CompadreData classes. 
In many cases the same method can be used for both classes, but some methods
are only valid for one class.

The metadata accessor function accesses the metadata of a CompadreData
object (it does not apply to CompadreM objects).

Metadata variables
These functions access certain variables from the metadata slot of a
CompadreData object. These methods are implemented using 
VariableName(CompadreDataObject) and are effectively the same as using 
`metadata(CompadreDataObject, "VariableName")`. Most variables are available 
for both compadre and comadre, but some are only available for either one or 
the other. Where a variable is not available for the chosen database, the 
function will simply return an error. The variables are: 
"SpeciesAuthor", "SpeciesAccepted", "CommonName", "Infraspecific" [comadre], 
"Genus" [compadre] or "GenusAccepted" [comadre], "GenusAuthor" [comadre],
"Family", "Order" , "Class", "Phylum", "Kingdom", 
"OrganismType", "DicotMonoc", [compadre], "AngioGymno" [compadre],
"Authors", "Journal", "YearPublication", "DOI.ISBN", "AdditionalSource",
"StudyDuration", "StudyStart", "StudyEnd", "AnnualPeriodicity",
"NumberPopulations", "MatrixCriteriaSize", "MatrixCriteriaOntogeny",
"MatrixCriteriaAge", "MatrixPopulation", 
"Lat", "Lon", "Altitude", "Country", "Continent", "Ecoregion",
"StudiedSex",
"MatrixComposite", "MatrixTreatment", "MatrixCaptivity",
"MatrixStartYear", "MatrixStartSeason", "MatrixStartMonth",
"MatrixEndYear", "MatrixEndSeason", "MatrixEndMonth",
"MatrixSplit", "MatrixFec", "Observation",
"MatrixDimension", "SurvivalIssue".
}
