% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreData.R
\docType{class}
\name{CompadreData}
\alias{CompadreData}
\alias{[,CompadreData,ANY,ANY,ANY-method}
\alias{metadata}
\alias{metadata,CompadreData-method}
\alias{SpeciesAuthor}
\alias{SpeciesAuthor,CompadreData-method}
\alias{SpeciesAccepted}
\alias{SpeciesAccepted,CompadreData-method}
\alias{CommonName}
\alias{CommonName,CompadreData-method}
\alias{Infraspecific}
\alias{Infraspecific,CompadreData-method}
\alias{Genus}
\alias{Genus,CompadreData-method}
\alias{GenusAccepted}
\alias{GenusAccepted,CompadreData-method}
\alias{GenusAuthor}
\alias{GenusAuthor,CompadreData-method}
\alias{Family}
\alias{Family,CompadreData-method}
\alias{Order}
\alias{Order,CompadreData-method}
\alias{Class}
\alias{Class,CompadreData-method}
\alias{Phylum}
\alias{Phylum,CompadreData-method}
\alias{Kingdom}
\alias{Kingdom,CompadreData-method}
\alias{OrganismType}
\alias{OrganismType,CompadreData-method}
\alias{DicotMonoc}
\alias{DicotMonoc,CompadreData-method}
\alias{AngioGymno}
\alias{AngioGymno,CompadreData-method}
\alias{Authors}
\alias{Authors,CompadreData-method}
\alias{Journal}
\alias{Journal,CompadreData-method}
\alias{YearPublication}
\alias{YearPublication,CompadreData-method}
\alias{DOI.ISBN}
\alias{DOI.ISBN,CompadreData-method}
\alias{AdditionalSource}
\alias{AdditionalSource,CompadreData-method}
\alias{StudyDuration}
\alias{StudyDuration,CompadreData-method}
\alias{StudyStart}
\alias{StudyStart,CompadreData-method}
\alias{StudyEnd}
\alias{StudyEnd,CompadreData-method}
\alias{AnnualPeriodicity}
\alias{AnnualPeriodicity,CompadreData-method}
\alias{NumberPopulations}
\alias{NumberPopulations,CompadreData-method}
\alias{MatrixCriteriaSize}
\alias{MatrixCriteriaSize,CompadreData-method}
\alias{MatrixCriteriaOntogeny}
\alias{MatrixCriteriaOntogeny,CompadreData-method}
\alias{MatrixCriteriaAge}
\alias{MatrixCriteriaAge,CompadreData-method}
\alias{MatrixPopulation}
\alias{MatrixPopulation,CompadreData-method}
\alias{Lat}
\alias{Lat,CompadreData-method}
\alias{Lon}
\alias{Lon,CompadreData-method}
\alias{Altitude}
\alias{Altitude,CompadreData-method}
\alias{Country}
\alias{Country,CompadreData-method}
\alias{Continent}
\alias{Continent,CompadreData-method}
\alias{Ecoregion}
\alias{Ecoregion,CompadreData-method}
\alias{StudiedSex}
\alias{StudiedSex,CompadreData-method}
\alias{MatrixComposite}
\alias{MatrixComposite,CompadreData-method}
\alias{MatrixTreatment}
\alias{MatrixTreatment,CompadreData-method}
\alias{MatrixCaptivity}
\alias{MatrixCaptivity,CompadreData-method}
\alias{MatrixStartYear}
\alias{MatrixStartYear,CompadreData-method}
\alias{MatrixStartSeason}
\alias{MatrixStartSeason,CompadreData-method}
\alias{MatrixStartMonth}
\alias{MatrixStartMonth,CompadreData-method}
\alias{MatrixEndYear}
\alias{MatrixEndYear,CompadreData-method}
\alias{MatrixEndSeason}
\alias{MatrixEndSeason,CompadreData-method}
\alias{MatrixEndMonth}
\alias{MatrixEndMonth,CompadreData-method}
\alias{MatrixSplit}
\alias{MatrixSplit,CompadreData-method}
\alias{MatrixFec}
\alias{MatrixFec,CompadreData-method}
\alias{Observation}
\alias{Observation,CompadreData-method}
\alias{MatrixDimension}
\alias{MatrixDimension,CompadreData-method}
\alias{SurvivalIssue}
\alias{SurvivalIssue,CompadreData-method}
\alias{mat}
\alias{mat,CompadreData-method}
\alias{Version}
\alias{Version,CompadreData-method}
\alias{DateCreated,CompadreData-method}
\alias{NumberAcceptedSpecies}
\alias{NumberAcceptedSpecies,CompadreData-method}
\alias{NumberStudies}
\alias{NumberStudies,CompadreData-method}
\alias{NumberMatrices}
\alias{NumberMatrices,CompadreData-method}
\title{CompadreData class}
\usage{
\S4method{[}{CompadreData,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

metadata(object, i, j)

\S4method{metadata}{CompadreData}(object, i)

SpeciesAuthor(object)

\S4method{SpeciesAuthor}{CompadreData}(object)

SpeciesAccepted(object)

\S4method{SpeciesAccepted}{CompadreData}(object)

CommonName(object)

\S4method{CommonName}{CompadreData}(object)

Infraspecific(object)

\S4method{Infraspecific}{CompadreData}(object)

Genus(object)

\S4method{Genus}{CompadreData}(object)

GenusAccepted(object)

\S4method{GenusAccepted}{CompadreData}(object)

GenusAuthor(object)

\S4method{GenusAuthor}{CompadreData}(object)

Family(object)

\S4method{Family}{CompadreData}(object)

Order(object)

\S4method{Order}{CompadreData}(object)

Class(object)

\S4method{Class}{CompadreData}(object)

Phylum(object)

\S4method{Phylum}{CompadreData}(object)

Kingdom(object)

\S4method{Kingdom}{CompadreData}(object)

OrganismType(object)

\S4method{OrganismType}{CompadreData}(object)

DicotMonoc(object)

\S4method{DicotMonoc}{CompadreData}(object)

AngioGymno(object)

\S4method{AngioGymno}{CompadreData}(object)

Authors(object)

\S4method{Authors}{CompadreData}(object)

Journal(object)

\S4method{Journal}{CompadreData}(object)

YearPublication(object)

\S4method{YearPublication}{CompadreData}(object)

DOI.ISBN(object)

\S4method{DOI.ISBN}{CompadreData}(object)

AdditionalSource(object)

\S4method{AdditionalSource}{CompadreData}(object)

StudyDuration(object)

\S4method{StudyDuration}{CompadreData}(object)

StudyStart(object)

\S4method{StudyStart}{CompadreData}(object)

StudyEnd(object)

\S4method{StudyEnd}{CompadreData}(object)

AnnualPeriodicity(object)

\S4method{AnnualPeriodicity}{CompadreData}(object)

NumberPopulations(object)

\S4method{NumberPopulations}{CompadreData}(object)

MatrixCriteriaSize(object)

\S4method{MatrixCriteriaSize}{CompadreData}(object)

MatrixCriteriaOntogeny(object)

\S4method{MatrixCriteriaOntogeny}{CompadreData}(object)

MatrixCriteriaAge(object)

\S4method{MatrixCriteriaAge}{CompadreData}(object)

MatrixPopulation(object)

\S4method{MatrixPopulation}{CompadreData}(object)

Lat(object)

\S4method{Lat}{CompadreData}(object)

Lon(object)

\S4method{Lon}{CompadreData}(object)

Altitude(object)

\S4method{Altitude}{CompadreData}(object)

Country(object)

\S4method{Country}{CompadreData}(object)

Continent(object)

\S4method{Continent}{CompadreData}(object)

Ecoregion(object)

\S4method{Ecoregion}{CompadreData}(object)

StudiedSex(object)

\S4method{StudiedSex}{CompadreData}(object)

MatrixComposite(object)

\S4method{MatrixComposite}{CompadreData}(object)

MatrixTreatment(object)

\S4method{MatrixTreatment}{CompadreData}(object)

MatrixCaptivity(object)

\S4method{MatrixCaptivity}{CompadreData}(object)

MatrixStartYear(object)

\S4method{MatrixStartYear}{CompadreData}(object)

MatrixStartSeason(object)

\S4method{MatrixStartSeason}{CompadreData}(object)

MatrixStartMonth(object)

\S4method{MatrixStartMonth}{CompadreData}(object)

MatrixEndYear(object)

\S4method{MatrixEndYear}{CompadreData}(object)

MatrixEndSeason(object)

\S4method{MatrixEndSeason}{CompadreData}(object)

MatrixEndMonth(object)

\S4method{MatrixEndMonth}{CompadreData}(object)

MatrixSplit(object)

\S4method{MatrixSplit}{CompadreData}(object)

MatrixFec(object)

\S4method{MatrixFec}{CompadreData}(object)

Observation(object)

\S4method{Observation}{CompadreData}(object)

MatrixDimension(object)

\S4method{MatrixDimension}{CompadreData}(object)

SurvivalIssue(object)

\S4method{SurvivalIssue}{CompadreData}(object)

mat(object, ...)

\S4method{mat}{CompadreData}(object)

Version(object, ...)

\S4method{Version}{CompadreData}(object)

\S4method{DateCreated}{CompadreData}(object)

NumberAcceptedSpecies(object, ...)

\S4method{NumberAcceptedSpecies}{CompadreData}(object)

NumberStudies(object, ...)

\S4method{NumberStudies}{CompadreData}(object)

NumberMatrices(object, ...)

\S4method{NumberMatrices}{CompadreData}(object)
}
\description{
Needs description

This method enables subsetting the data using square brackets, i.e. for a
CompadreData object "DB" one can use DB[1:2, 1:10] to get a new CompadreData
object that includes the first two matrices and the the first 10 columns of 
their associated metadata. It's possible to pass numeric vectors, logical 
vectors, or character vectors that match the row or column names of the 
metadata. The function calls subsetDB to do the work of subsetting rows.

The metadata accessor function accesses the metadata of a CompadreData
object (it does not apply to CompadreM objects).

The mat accessor function accesses the matrices of a CompadreData object
object.

The version (including subset if relevant) of a CompadreData object.

The number of accepted binary species names in a CompadreData object.

The number of different studies in a CompadreData object.

The number of CompadreM objects contained in a CompadreData object (i.e. the)
number of projection matrices).
}
