% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreData.R
\docType{class}
\name{CompadreDataClass}
\alias{CompadreDataClass}
\title{'CompadreData' class}
\description{
This page describes the 'CompadreData' class. A CompadreData object contains
many CompadreM objects, therefore many matrices. Methods for the CompadreData
class are mostly focused on extraction of relevant data. Each method includes 
a 'generic' function with the method implemented through that generic.
}
