% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreDB.R
\docType{class}
\name{CompadreDataMethods}
\alias{CompadreDataMethods}
\alias{asCompadreDB}
\alias{as.data.frame.CompadreDB}
\alias{as_tibble.CompadreDB}
\alias{head.CompadreDB}
\alias{merge.CompadreDB}
\alias{CompadreData}
\alias{CompadreData,CompadreDB-method}
\alias{$,CompadreDB-method}
\alias{$<-,CompadreDB-method}
\alias{names.CompadreDB}
\alias{VersionData}
\alias{VersionData,CompadreDB-method}
\alias{Version}
\alias{Version,CompadreDB-method}
\alias{DateCreated}
\alias{DateCreated,CompadreDB-method}
\alias{NumberAcceptedSpecies}
\alias{NumberAcceptedSpecies,CompadreDB-method}
\alias{NumberStudies}
\alias{NumberStudies,CompadreDB-method}
\alias{NumberMatrices}
\alias{NumberMatrices,CompadreDB-method}
\title{CompadreDB class}
\usage{
asCompadreDB(from)

\method{as.data.frame}{CompadreDB}(x, ...)

\method{as_tibble}{CompadreDB}(x)

\method{head}{CompadreDB}(x, n = 6L, ...)

\method{merge}{CompadreDB}(x, y, ...)

CompadreData(object)

\S4method{CompadreData}{CompadreDB}(object)

\S4method{$}{CompadreDB}(x, name)

\S4method{$}{CompadreDB}(x, name) <- value

\method{names}{CompadreDB}(x)

VersionData(object, ...)

\S4method{VersionData}{CompadreDB}(object)

Version(object, ...)

\S4method{Version}{CompadreDB}(object)

DateCreated(object, ...)

\S4method{DateCreated}{CompadreDB}(object)

NumberAcceptedSpecies(object, ...)

\S4method{NumberAcceptedSpecies}{CompadreDB}(object)

NumberStudies(object, ...)

\S4method{NumberStudies}{CompadreDB}(object)

NumberMatrices(object, ...)

\S4method{NumberMatrices}{CompadreDB}(object)
}
\arguments{
\item{from}{A legacy COM(P)ADRE database}

\item{x}{A CompadreDB object}

\item{...}{Ignored}

\item{n}{The number of rows to show}

\item{y}{A data.frame to merge with x}

\item{object}{A CompadreDB object}

\item{name}{The name of a column within x}

\item{value}{Vector of values to assign to the column}
}
\description{
This page describes methods for accessing any metadata information from 
CompadreDB objects.
}
