% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreDB-Tidyverse.R
\name{CompadreDB-Tidyverse}
\alias{CompadreDB-Tidyverse}
\alias{fortify.CompadreDB}
\title{Tidyverse methods for CompadreDB objects}
\usage{
fortify.CompadreDB(model, data, ...)
}
\arguments{
\item{model}{A CompadreDB object}

\item{data}{ignored}

\item{...}{ignored}
}
\description{
Tidyverse methods for CompadreDB objects
}
