% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreDB.R
\docType{class}
\name{CompadreDB}
\alias{CompadreDB}
\alias{asCompadreDB}
\alias{CompadreData}
\alias{CompadreData,CompadreDB-method}
\alias{$,CompadreDB-method}
\alias{$<-,CompadreDB-method}
\alias{VersionData}
\alias{VersionData,CompadreDB-method}
\alias{Version}
\alias{Version,CompadreDB-method}
\alias{DateCreated}
\alias{DateCreated,CompadreDB-method}
\title{CompadreDB Class}
\usage{
asCompadreDB(from)

CompadreData(object)

\S4method{CompadreData}{CompadreDB}(object)

\S4method{$}{CompadreDB}(x, name)

\S4method{$}{CompadreDB}(x, name) <- value

VersionData(object)

\S4method{VersionData}{CompadreDB}(object)

Version(object)

\S4method{Version}{CompadreDB}(object)

DateCreated(object)

\S4method{DateCreated}{CompadreDB}(object)
}
\arguments{
\item{from}{A legacy COM(P)ADRE database}

\item{object}{A CompadreDB object}

\item{x}{A CompadreDB object}

\item{name}{The name of a column within x}

\item{value}{Vector of values to assign to the column}
}
\description{
This page describes the CompadreDB class, including methods for accessing the
various components of CompadreDB objects.
}
\section{Slots}{

\describe{
\item{\code{CompadreData}}{A tibble-style data frame with a list-column of matrix
population models (column \code{mat}) and a variety of other metadata
columns.}

\item{\code{VersionData}}{A list with elements \code{Version} (database version
number), \code{DateCreated} (date of version release), and \code{Agreement}
(a url link to the User Agreement)}
}}

