% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreDB.R
\docType{class}
\name{CompadreDBMethods}
\alias{CompadreDBMethods}
\alias{asCompadreDB}
\alias{[,CompadreDB,ANY,ANY,ANY-method}
\title{CompadreDB class}
\usage{
asCompadreDB(from)

\S4method{[}{CompadreDB,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\description{
This page describes methods for accessing any metadata information from 
CompadreDB objects.
}
