% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreM.R
\docType{class}
\name{CompadreMClass}
\alias{CompadreMClass}
\title{CompadreM class}
\description{
This page describes the CompadreM class. A CompadreM object describes
a single matrix and its associated data. Methods for the CompadreM class
are mostly focused on extraction of relevant data. Each method includes 
a 'generic' function with the method implemented through that generic.
}
