% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreData.R, R/CompadreM.R
\docType{methods}
\name{CompadreDataMethods}
\alias{CompadreDataMethods}
\alias{CompadreMMethods}
\alias{matA,CompadreM-method}
\title{Methods for 'CompadreData' class}
\usage{
matAlist(object)

matA(object)

\S4method{matA}{CompadreM}(object)
}
\description{
Below are methods for the CompadreData class. These are mostly focused 
on extraction of relevant data. Each method includes a 'generic' 
function with the implemented through that generic.

Below are methods for the CompadreM class. These are mostly focused 
on extraction of relevant data. Each method includes a 'generic' 
function with the implemented through that generic.
}
