% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassUnionMethods.R
\docType{class}
\name{CompadreMorData-class}
\alias{CompadreMorData-class}
\alias{matA}
\alias{matA,CompadreMorData-method}
\alias{matU}
\alias{matU,CompadreMorData-method}
\alias{matF}
\alias{matF,CompadreMorData-method}
\alias{matC}
\alias{matC,CompadreMorData-method}
\alias{matrixClass}
\alias{matrixClass,CompadreMorData-method}
\alias{stageNames,CompadreMorData-method}
\title{Accessor methods for CompadreM and CompadreData objects}
\usage{
matA(object)

\S4method{matA}{CompadreMorData}(object)

matU(object)

\S4method{matU}{CompadreMorData}(object)

matF(object)

\S4method{matF}{CompadreMorData}(object)

matC(object)

\S4method{matC}{CompadreMorData}(object)

matrixClass(object)

\S4method{matrixClass}{CompadreMorData}(object)

\S4method{stageNames}{CompadreMorData}(object)
}
\description{
This page describes methods for both the CompadreM and CompadreData classes. 
In many cases the same method can be used for both classes, but some methods
are only valid for one class.
}
