% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDB.R
\name{convertLegacyDB}
\alias{convertLegacyDB}
\title{Convert legacy COM(P)ADRE database to CompadreDB object}
\usage{
convertLegacyDB(db)
}
\arguments{
\item{db}{The legacy COM(P)ADRE database to convert}
}
\value{
A \code{CompadreDB} object
}
\description{
Utility to convert older versions of database to
modern class
}
