% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDB.R
\name{fetchDB}
\alias{fetchDB}
\title{Fetch the COM(P)ADRE database from compadre-db.org or a local file}
\usage{
fetchDB(db)
}
\arguments{
\item{db}{Either "comadre" or "compadre" (case insensitive) to fetch the most
recent database from \url{https://compadre-db.org}, or a path to an
existing COMPADRE database (i.e. .RData file) stored on the local machine.}
}
\value{
A CompadreDB object
}
\description{
Fetch the current version of the COM(P)ADRE database from
\url{https://compadre-db.org}, or load any version stored in a local .RData
file.
}
\examples{
\dontrun{
compadre <- fetchDB('compadre')
}
}
\author{
Patrick M. Barks <patrick.barks@gmail.com>
}
