% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkspecies.R
\name{findSpecies}
\alias{findSpecies}
\title{Utility function for checkSpecies}
\usage{
findSpecies(x)
}
\arguments{
\item{x}{A character vector of species names}
}
\value{
A logical indicating whether the species name is in the 
COM(P)ADRE object
}
\description{
Utility function for checkSpecies
}
