% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanMatF.R
\name{getMeanMatF}
\alias{getMeanMatF}
\title{Calculates a population-specific grand mean fecundity matrix for each set of
matrices in a COM(P)ADRE database object}
\usage{
getMeanMatF(db)
}
\arguments{
\item{db}{A COM(P)ADRE database object.}
}
\value{
Returns a list which contains the mean fecundity matrix associated
  with a given row of the database, or NA if there is only a single matrix
  from the relevant population within the db.
}
\description{
This function takes a COM(P)ADRE database object and calculates a grand mean
fecundity matrix for each unique population (a mean of all
population-specific fecundity matrices, including fecundity matrices for
which \code{MatrixComposite == 'Mean'}). Here, a unique study population is
defined as a unique combination of the metadata columns Authors,
YearPublication, DOI.ISBN, SpeciesAuthor, MatrixPopulation, and
MatrixDimension. The main purpose of this function is to identify stage
classes that are \emph{potentially} reproductive (i.e. the absense of
fecundity in a given stage class and year does not necessarily indicate that
the stage in question is non-reproductive).
}
\examples{
\dontrun{
# print set of matrices (A, U, F, C) associated with row 2 of database
compadre$mat[[2]]

# create list of meanMatFs
meanF <- getMeanMatF(compadre)

# print meanMatF associated with row 2 of database
compadre_with_meanF$mat[[2]]
}
}
\author{
Danny Buss <dlb50@cam.ac.uk>

Julia Jones <juliajones@biology.sdu.dk>

Iain Stott <stott@biolgy.ox.ac.uk>

Patrick Barks <patrick.barks@gmail.com>
}
