% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyReprodStages.R
\name{identifyReproStages}
\alias{identifyReproStages}
\title{Identify reproductive stages}
\usage{
identifyReproStages(matF)
}
\arguments{
\item{matF}{fecundity matrix}
}
\value{
A numeric vector with zeros corresponding to non-reproductive stages
and ones corresponding to reproductive stages
}
\description{
Takes a fecundity matrix and returns a vector of reproductive and non-
reproductive stages. This is a preparatory step to collapsing the matrix
model into combined propagule, pre-reproductive, reproductive and
post-reproductive stages.
}
\examples{
matF <- matrix(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
                 0, 0, 0, 0, 1.0409262, 0.5040727, 0.016433, 0),
                 nrow = 5, byrow = FALSE)
identifyReproStages(matF)
}
\keyword{~~}
