% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDBs.R
\name{mergeDBs}
\alias{mergeDBs}
\title{Merge two Compadre/Comadre data sets together}
\usage{
mergeDBs(db1, db2)
}
\arguments{
\item{db1}{A \code{CompadreDB} object or an older version
of \code{COM(P)ADRE} in a \code{list}.}

\item{db2}{A \code{CompadreDB} object or an older version
of \code{COM(P)ADRE} in a \code{list}.}
}
\value{
A \code{CompadreDB} object containing both
databases.
}
\description{
Merges two data set objects together. These can either
be S4 objects of class \code{CompadreDB}, S3 list objects that 
contain an older version of the database, or a combination of the
two.
}
\examples{
\dontrun{
CompadreData(Compadre)
CompadreData(Comadre)

BigDB <- mergeDBs(Compadre, Comadre)

}

}
\author{
Sam Levin
}
