% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm_mean.R
\name{mpm_mean}
\alias{mpm_mean}
\alias{mat_mean}
\title{Calculate a mean over a list of matrices or CompadreMat objects}
\usage{
mat_mean(x)

mpm_mean(x)
}
\arguments{
\item{x}{List of matrices (\code{mat_mean}) or list of CompadreMat objects
(\code{mpm_mean}), all of same dimension}
}
\value{
A matrix (\code{mat_mean}) or a CompadreMat object (\code{mpm_mean}).
}
\description{
Calculates an element-wise mean over a list of matrices or CompadreMat
objects of constant dimension.

Does not currently have an option to remove \code{NA}.
}
\examples{
# there are four rows for species 'Haplopappus_radiatus' in Compadre
mpms <- Compadre$mat[Compadre$SpeciesAuthor == "Haplopappus_radiatus"]
mpm_mean(mpms)

# extract list of matA and take mean
mats <- matA(mpms)
mat_mean(mats)
}
\author{
Patrick Barks <patrick.barks@gmail.com>
}
