% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm_summaries.R
\name{mpm_median}
\alias{mpm_median}
\title{Calculate a standard deviation over a list of matrices}
\usage{
mpm_median(x, na.rm = FALSE)
}
\arguments{
\item{x}{List of matrices all of same dimension}

\item{na.rm}{Logical indicating whether missing values should be excluded
(see \emph{Details}). Defaults to \code{FALSE}.}
}
\value{
A matrix containing the standard deviation of each element across all
  matrices in the list
}
\description{
Calculates an element-wise standard deviation over a list of matrices of
constant dimension.
}
\details{
If \code{na.rm == TRUE}, missing values are ignored in the calculation of the
mean matrix. If \code{na.rm == TRUE} and a given element is \code{NA} in
\emph{every} matrix within \code{x}, the value returned for that element will
be \code{0}.
}
\seealso{
Other data management: 
\code{\link{cdb_flatten}()},
\code{\link{cdb_id_stages}()},
\code{\link{cdb_id_studies}()},
\code{\link{cdb_id}()},
\code{\link{cdb_mean_matF}()},
\code{\link{cdb_rbind}()},
\code{\link{cdb_unflatten}()},
\code{\link{cdb_unnest}()},
\code{\link{mpm_mean}()},
\code{\link{mpm_sd}()},
\code{\link{string_representation}}
}
\author{
Darren Norris
}
\concept{data management}
