% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcompadre-package.R
\docType{package}
\name{Rcompadre-package}
\alias{Rcompadre}
\alias{Rcompadre-package}
\title{Rcompadre: Utilities for using the 'COM(P)ADRE' Matrix Model Database}
\description{
Utility functions for interacting with the 'COMPADRE' and 'COMADRE' databases of matrix population models. Described in Jones et al. (2021) \doi{10.1101/2021.04.26.441330}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonesor/Rcompadre}
  \item Report bugs at \url{https://github.com/jonesor/Rcompadre/issues}
}

}
\author{
\strong{Maintainer}: Owen Jones \email{jones@biology.sdu.dk} (\href{https://orcid.org/0000-0001-5720-4686}{ORCID})

Authors:
\itemize{
  \item Patrick Barks \email{barks@biology.sdu.dk} (\href{https://orcid.org/0000-0002-5947-8151}{ORCID})
  \item Danny Buss \email{dlb50@cam.ac.uk}
  \item Roberto Salguero-Gomez \email{rob.salguero@zoo.ox.ac.uk} (\href{https://orcid.org/0000-0002-6085-4433}{ORCID})
  \item Iain Stott \email{stott@biology.sdu.dk} (\href{https://orcid.org/0000-0003-2724-7436}{ORCID})
  \item William K. Petry \email{wpetry@ncsu.edu} (\href{https://orcid.org/0000-0002-5230-5987}{ORCID})
  \item Tamora James \email{tamoradjames@protonmail.com} (\href{https://orcid.org/0000-0003-1363-4742}{ORCID})
  \item Julia Jones \email{juagjones@gmail.com} (\href{https://orcid.org/0000-0001-9223-1778}{ORCID})
  \item Gesa Römer \email{gesa@biology.sdu.dk} (\href{https://orcid.org/0000-0002-4859-5870}{ORCID})
  \item Sam Levin \email{levisc8@gmail.com} (\href{https://orcid.org/0000-0002-3289-9925}{ORCID})
}

}
\keyword{internal}
