% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangeMatrix.R
\name{rearrangeMatrix}
\alias{rearrangeMatrix}
\title{Rearrange matrix stages to segregate reproductive and non-reproductive stages}
\usage{
rearrangeMatrix(matU, matF, repro_stages)
}
\arguments{
\item{matU}{Survival matrix}

\item{matF}{Fecundity matrix}

\item{repro_stages}{Logical vector identifying which stages reproductive}
}
\value{
Returns a list with 5 elements: the rearranged survival matrix
  (\code{matU}), the rearranged fecundity matrix (\code{matF}), the
  rearranged vector of reproductive stages (\code{repro_stages}), the numeric
  index for any rearranged inter-reproductive stages (\code{nonRepInterRep}),
  and the numeric index for the maximum reproductive stage in the rearranged
  reproductive stage vector (\code{maxRep}).
}
\description{
Rearrange matrix stages so that all inter-reproductive stages fall in the
final rows/columns of the matrix. This is a preparatory step to collapsing
the matrix model into a standardized set of stages (e.g. propagule,
pre-reproductive, reproductive, and post-reproductive).
}
\examples{
matU <- rbind(c(0, 0, 0, 0, 0), c(0.1, 0.16, 0, 0, 0), c(0.2, 0.23, 0.12, 0,
0), c(0, 0, 0.34, 0.53, 0), c(0, 0, 0, 0.34, 0))

matF <- rbind(c(0, 0, 0, 0, 0), c(0, 0.2, 0, 0.1, 0), c(0, 0.2, 0, 0.1, 0),
c(0, 0, 0, 0, 0), c(0, 0, 0, 0, 0))

repro_stages <- c(0, 1, 0, 1, 0)
rearrangeMatrix(matU, matF, repro_stages)
}
\author{
Rob Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>
}
