% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassUnionMethods.R
\name{stageStatus}
\alias{stageStatus}
\title{The 'matrixClassAuthor' function extracts the matrixClassAuthor column from 
the matrixClass data frame from a CompadreMat or CompadreDB object. 
For CompadreMat objects, this is a single character vector, for CompadreDB objects 
this is a list of character vectors. The matrixClassAuthor data
describes the names of the stages as determined by the author of the original 
work the matrix was sourced from.}
\usage{
stageStatus(object)
}
\description{
The 'matrixClassAuthor' function extracts the matrixClassAuthor column from 
the matrixClass data frame from a CompadreMat or CompadreDB object. 
For CompadreMat objects, this is a single character vector, for CompadreDB objects 
this is a list of character vectors. The matrixClassAuthor data
describes the names of the stages as determined by the author of the original 
work the matrix was sourced from.
}
