% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtomatrix.R
\name{stringToMatrix}
\alias{stringToMatrix}
\title{Converts a matrix in character string format to numeric matrix format}
\usage{
stringToMatrix(A)
}
\arguments{
\item{A}{A square matrix in the form of a string, begining with an open
square bracket and ending with a closed square bracket, with individual
matrix cell entries separated by a space.}
}
\value{
A square numeric \code{matrix}.
}
\description{
This function converts a matrix in character string format (e.g. "[0.2 0.3
0.1 0]") to a square numeric matrix.
}
\examples{
x1 <- "[3.3 5.2 6.1 0.1 NA 0.3 0.2 0.4 0.1]"
stringToMatrix(x1)

x2 <- "[0 0 0 0.42 0.65 0 0 0 0 0.52 0 0 0 0 0.2 0]"
stringToMatrix(x2)

\dontrun{
# non-square matrix
x3 <- "[0.42 0.52 0.15 0.23 0.14]"
stringToMatrix(x3)
}

}
\seealso{
DBToFlat
}
\author{
Owen R. Jones <jones@biology.sdu.dk>
}
