% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\docType{methods}
\name{Subset-CompadreDB}
\alias{Subset-CompadreDB}
\alias{[,CompadreDB,ANY,ANY,ANY-method}
\alias{subset.CompadreDB}
\title{Subsetting CompadreDB objects}
\usage{
\S4method{[}{CompadreDB,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\method{subset}{CompadreDB}(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x}{A CompadreDB object}

\item{i}{row indices (see \link{[.data.frame})}

\item{j}{column indices (see \link{[.data.frame})}

\item{...}{ignored}

\item{drop}{ignored}

\item{subset}{logical expression indicating which rows to keep}

\item{select}{expression indicating which columns to keep}
}
\description{
This method enables subsetting the data using square brackets, i.e. for a
CompadreDB object "DB" one can use DB[1:2, 1:10] to get a new CompadreDB
object that includes the first two matrices and the the first 10 columns of 
their associated metadata. It's possible to pass logical vectors, numeric
vectors, or character vectors that match the row or column names of the 
metadata.
}
