% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetDB.R
\name{subsetDB}
\alias{subsetDB}
\title{Subsets the COMPADRE/COMADRE database}
\usage{
subsetDB(db, sub)
}
\arguments{
\item{db}{The COMPADRE or COMADRE database object.}

\item{sub}{An argument made using logical operators (see `subset`) with
which to subset the data base. Any of the variables contained in the
metadata part of the COMPADRE/COMADRE database may be used.}
}
\value{
Returns a subset of the database, with the same structure, but where
the records in the metadata match the criteria given in the `sub` argument.
}
\description{
Subset the COMPADRE/COMADRE database by logical argument.
}
\examples{
\dontrun{
ssData <- subsetDB(compadre, MatrixDimension > 3)
ssData <- subsetDB(compadre, MatrixDimension > 3 & MatrixComposite == "Mean")
ssData <- subsetDB(comadre, Continent == "Africa" & Class == "Mammalia")
ssData <- subsetDB(comadre, Altitude > 1000 & Altitude < 1500)
}
}
\author{
Owen R. Jones <jones@biology.sdu.dk>

Rob Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Bruce Kendall <kendall@bren.ucsb.edu>
}
