% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendMeanMatF.R
\name{appendMeanMatF}
\alias{appendMeanMatF}
\title{Appends a population-specific mean fecundity matrix to each set of matrices
in a COM(P)ADRE database object}
\usage{
appendMeanMatF(db)
}
\arguments{
\item{db}{A COM(P)ADRE database object.}
}
\value{
Returns db, but with an extra entry in slot `mat` which contains the
  mean fecundity matrix associated with a given row of the database, or NA if
  there are no mean matrices within the db from the relevant population.
}
\description{
This function takes a COM(P)ADRE database object, calculates a mean fecundity
matrix for each unique population (a mean of all population-specific
fecundity matrices for which MatrixComposite == "Mean"), and appends the
relevant mean fecundity matrix to each original set of matrices in the
database. For the purpose of this function, a unique study population is
defined as a unique combination of the metadata columns Authors,
YearPublication, DOI.ISBN, SpeciesAuthor, MatrixPopulation, and
MatrixDimension.
}
\examples{
\dontrun{
# print set of matrices (A, U, F, C) associated with row 2 of database
compadre$mat[[2]]

# append meanMatFs to database object
compadre_with_meanF <- appendMeanMatF(compadre)

# print matrices associated with row 2 of database, including meanMatF
compadre_with_meanF$mat[[2]]
}
}
\author{
Danny Buss <dlb50@cam.ac.uk>

Julia Jones <juliajones@biology.sdu.dk>

Iain Stott <stott@biolgy.ox.ac.uk>

Patrick Barks <patrick.barks@gmail.com>
}
