% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDB.R
\name{cdb_flag}
\alias{cdb_flag}
\title{Flag potential issues in matrices of a COM(P)ADRE database}
\usage{
cdb_flag(cdb)
}
\arguments{
\item{cdb}{A CompadreDB object}
}
\value{
Returns \code{cdb} with extra columns appended to the data slot to
  indicate (TRUE/FALSE) whether there are potential problems with the
  matrices corresponding to a given row of the data, including whether matA
  is ergodic, primitive, and irreducible, and whether matU is singular (i.e.
  cannot be inverted).
}
\description{
Adds columns to the data slot of a CompadreDB object that flag potential
problems in the matrix population models, such as matrices containing missing
values, or projection models that are non-ergodic, reducible, or imprimitive.
These columns can subsequently be used to subset the database by logical
argument.
}
\examples{
CompadreFlag <- cdb_flag(Compadre)

}
\author{
Julia Jones <juliajones@biology.sdu.dk>

Roberto Salguero-Goméz <rob.salguero@zoo.ox.ac.uk>

Danny Buss <dlb50@cam.ac.uk>

Patrick Barks <patrick.barks@gmail.com>
}
