% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_merge.R
\name{cdb_rbind}
\alias{cdb_rbind}
\title{Merge two COM(P)ADRE databases via row-bind}
\usage{
cdb_rbind(cdb1, cdb2)
}
\arguments{
\item{cdb1, cdb2}{CompadreDB objects}
}
\value{
A CompadreDB object created by binding the rows of \code{cdb1} and
  \code{cdb2}
}
\description{
Merge two CompadreDB objects via a row-bind of the data slots.
}
\examples{
Compadre1 <- subset(Compadre, Continent == "Asia")
Compadre2 <- subset(Compadre, Continent == "Africa")

cdb_rbind(Compadre1, Compadre2)

}
\author{
Sam Levin
}
