% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_subset_wild.R
\name{cdb_subset_wild}
\alias{cdb_subset_wild}
\title{Subset a COM(P)ADRE database to wild, unmanipulated populations}
\usage{
cdb_subset_wild(cdb)
}
\arguments{
\item{cdb}{A CompadreDB object}
}
\value{
A CompadreDB object reflecting the subset of \code{cdb} corresponding
  to wild (\code{MatrixCaptivity = "W"}), unmanipulated
  (\code{MatrixTreatment == "Unmanipulated"}) populations.
}
\description{
Subsets a CompadreDB object to the rows reflecting wild, unmanipulated
populations \code{(MatrixCaptivity == "W" & MatrixTreatment ==
"Unmanipulated")}.

Note that such a subset may be overly stringent for some analyses, as it
excludes populations subject to "treatments" that may not necessitate
exclusion (e.g. "treatment" may reflect grazing status, time since fire,
etc.).
}
\examples{
CompadreWild <- cdb_subset_wild(Compadre)

}
\author{
Patrick M. Barks <patrick.barks@gmail.com>
}
