% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDatabase.R
\name{cleanDatabase}
\alias{cleanDatabase}
\title{Flag potential issues in matrices of a COM(P)ADRE database.}
\usage{
cleanDatabase(db)
}
\arguments{
\item{db}{A COM(P)ADRE database object.}
}
\value{
Returns db with extra columns appended to the metadata to indicate
  (TRUE/FALSE) whether there are potential problems with the matrices
  corresponding to a given row of the metadata, including whether matA is
  ergodic, primitive, and irreducible.
}
\description{
This function adds columns to the metatadata slot of a COM(P)ADRE database
object that flag potential problems in the matrices, such as when matrices
contain missing values. These columns can subsequently can be used to subset
the COM(P)ADRE database by logical argument.
}
\examples{
\dontrun{
compadre_clean <- cleanDatabase(compadre)
}

}
\author{
Julia Jones <juliajones@biology.sdu.dk>
Roberto Salguero-Goméz
Danny Buss
Patrick Barks
}
\keyword{utilities}
