% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreDB.R
\docType{methods}
\name{CompadreDataMethods}
\alias{CompadreDataMethods}
\alias{as.data.frame.CompadreDB}
\alias{as_tibble.CompadreDB}
\alias{head.CompadreDB}
\alias{merge.CompadreDB}
\alias{subset.CompadreDB}
\alias{CompadreData,CompadreDB-method}
\alias{$,CompadreDB-method}
\alias{$<-,CompadreDB-method}
\alias{names.CompadreDB}
\alias{VersionData}
\alias{VersionData,CompadreDB-method}
\alias{Version}
\alias{Version,CompadreDB-method}
\alias{DateCreated,CompadreDB-method}
\alias{NumberAcceptedSpecies}
\alias{NumberAcceptedSpecies,CompadreDB-method}
\alias{NumberStudies}
\alias{NumberStudies,CompadreDB-method}
\alias{NumberMatrices}
\alias{NumberMatrices,CompadreDB-method}
\title{The CompadreData accessor function accesses the data of a CompadreDB
object, which is a data frame including the matrices and all the 
metadata variables. Further methods described below allow the user to access 
individual metadata variables as vectors.}
\usage{
\method{as.data.frame}{CompadreDB}(x)

\method{as_tibble}{CompadreDB}(x)

\method{head}{CompadreDB}(x)

\method{merge}{CompadreDB}(x, y, ...)

\method{subset}{CompadreDB}(x, subset, select, drop = FALSE, ...)

CompadreData(object)

\S4method{CompadreData}{CompadreDB}(object)

\S4method{$}{CompadreDB}(x, name)

\S4method{$}{CompadreDB}(x, name) <- value

\method{names}{CompadreDB}(x)

VersionData(object, ...)

\S4method{VersionData}{CompadreDB}(object)

Version(object, ...)

\S4method{Version}{CompadreDB}(object)

\S4method{DateCreated}{CompadreDB}(object)

NumberAcceptedSpecies(object, ...)

\S4method{NumberAcceptedSpecies}{CompadreDB}(object)

NumberStudies(object, ...)

\S4method{NumberStudies}{CompadreDB}(object)

NumberMatrices(object, ...)

\S4method{NumberMatrices}{CompadreDB}(object)
}
\description{
The CompadreData accessor function accesses the data of a CompadreDB
object, which is a data frame including the matrices and all the 
metadata variables. Further methods described below allow the user to access 
individual metadata variables as vectors.

These functions access certain variables from the metadata slot of a
CompadreDB object. These methods are implemented using 
CompadreDBObject$VariableName and are effectively the same as using 
`CompadreData(CompadreDBObject, "VariableName")`. Most variables are available 
for both compadre and comadre, but some are only available for either one or 
the other. Where a variable is not available for the chosen database, the 
function will simply return an error. The variables are: 
"SpeciesAuthor", "SpeciesAccepted", "CommonName", "Infraspecific" [comadre], 
"Genus" [compadre] or "GenusAccepted" [comadre], "GenusAuthor" [comadre],
"Family", "Order" , "Class", "Phylum", "Kingdom", 
"OrganismType", "DicotMonoc", [compadre], "AngioGymno" [compadre],
"Authors", "Journal", "YearPublication", "DOI.ISBN", "AdditionalSource",
"StudyDuration", "StudyStart", "StudyEnd", "AnnualPeriodicity",
"NumberPopulations", "MatrixCriteriaSize", "MatrixCriteriaOntogeny",
"MatrixCriteriaAge", "MatrixPopulation", 
"Lat", "Lon", "Altitude", "Country", "Continent", "Ecoregion",
"StudiedSex",
"MatrixComposite", "MatrixTreatment", "MatrixCaptivity",
"MatrixStartYear", "MatrixStartSeason", "MatrixStartMonth",
"MatrixEndYear", "MatrixEndSeason", "MatrixEndMonth",
"MatrixSplit", "MatrixFec", "Observation",
"MatrixDimension", "SurvivalIssue".

All version information (including subset information) for a CompadreDB 
object, as a list.

The version (including subset if relevant) of a CompadreDB object.

The number of accepted binary species names in a CompadreDB object.

The number of different studies in a CompadreDB object.

The number of CompadreMat objects contained in a CompadreDB object (i.e. the)
number of projection matrices).
}
