% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreDB-Tidyverse.R
\name{CompadreDB-Tidyverse}
\alias{CompadreDB-Tidyverse}
\alias{fortify.CompadreDB}
\alias{filter.CompadreDB}
\alias{slice.CompadreDB}
\alias{arrange.CompadreDB}
\alias{mutate.CompadreDB}
\alias{group_by.CompadreDB}
\alias{ungroup.CompadreDB}
\alias{summarize.CompadreDB}
\alias{summarise.CompadreDB}
\alias{select.CompadreDB}
\alias{rename.CompadreDB}
\alias{left_join.CompadreDB}
\alias{right_join.CompadreDB}
\alias{inner_join.CompadreDB}
\alias{full_join.CompadreDB}
\title{Tidyverse methods for CompadreDB objects}
\usage{
\method{fortify}{CompadreDB}(model, data, ...)

\method{filter}{CompadreDB}(.data, ...)

\method{slice}{CompadreDB}(.data, ...)

\method{arrange}{CompadreDB}(.data, ...)

\method{mutate}{CompadreDB}(.data, ...)

\method{group_by}{CompadreDB}(.data, ..., add = FALSE)

\method{ungroup}{CompadreDB}(x, ...)

\method{summarize}{CompadreDB}(.data, ...)

\method{summarise}{CompadreDB}(.data, ...)

\method{select}{CompadreDB}(.data, ...)

\method{rename}{CompadreDB}(.data, ...)

\method{left_join}{CompadreDB}(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

\method{right_join}{CompadreDB}(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

\method{inner_join}{CompadreDB}(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

\method{full_join}{CompadreDB}(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)
}
\arguments{
\item{data}{see \code{\link[ggplot2]{fortify}}}

\item{...}{other arguments}

\item{add}{Logical indicating whether to overwrite existing groups
(\code{FALSE}) or add to any existing groups (\code{TRUE})}

\item{x, model, .data}{A CompadreDB object}

\item{y}{see \code{\link[dplyr]{join}}}

\item{by}{see \code{\link[dplyr]{join}}}

\item{copy}{see \code{\link[dplyr]{join}}}

\item{suffix}{see \code{\link[dplyr]{join}}}
}
\value{
No return value, called for side effects
}
\description{
CompadreDB methods for functions in \link[dplyr]{dplyr} and
\link[ggplot2]{ggplot2}.
}
