% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreMat.R, R/ClassUnionMethods.R
\docType{class}
\name{CompadreMatrixMethods}
\alias{CompadreMatrixMethods}
\alias{CompadreMatOrDB-class}
\alias{matA}
\alias{matA,CompadreMatOrDB-method}
\alias{matU}
\alias{matU,CompadreMatOrDB-method}
\alias{matF}
\alias{matF,CompadreMatOrDB-method}
\alias{matC}
\alias{matC,CompadreMatOrDB-method}
\alias{matrixClass}
\alias{matrixClass,CompadreMatOrDB-method}
\alias{stageNames,CompadreMatOrDB-method}
\alias{MatrixClassAuthor,CompadreMatOrDB-method}
\alias{stageStatus,CompadreMatOrDB-method}
\alias{MatrixClassOrganized,CompadreMatOrDB-method}
\title{Methods for working with matrices in com(p)adre}
\usage{
matA(object)

\S4method{matA}{CompadreMatOrDB}(object)

matU(object)

\S4method{matU}{CompadreMatOrDB}(object)

matF(object)

\S4method{matF}{CompadreMatOrDB}(object)

matC(object)

\S4method{matC}{CompadreMatOrDB}(object)

matrixClass(object)

\S4method{matrixClass}{CompadreMatOrDB}(object)

\S4method{stageNames}{CompadreMatOrDB}(object)

\S4method{MatrixClassAuthor}{CompadreMatOrDB}(object)

\S4method{stageStatus}{CompadreMatOrDB}(object)

\S4method{MatrixClassOrganized}{CompadreMatOrDB}(object)
}
\description{
This page describes methods for accessing any matrix information from 
CompadreMat and CompadreDB objects.

Most methods for working with matrices are applicable to both CompadreMat and 
CompadreDB objects. These are described on this page (along with a couple) 
of methods applicable to only CompadreMat or CompadreDB objects).
}
