% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDBs.R
\name{compareDBs}
\alias{compareDBs}
\alias{dbCompare}
\title{A function to compare two COMPADRE/COMADRE database versions}
\usage{
compareDBs(db1, db2, verbose = FALSE)

dbCompare(db1, db2, verbose = FALSE)
}
\arguments{
\item{db1, db2}{COM(P)ADRE database objects to compare.}

\item{verbose}{A logical argument indicating whether or not to return lots
of detail.}
}
\value{
Prints a summary to the screen of the comparison between two databases.
}
\description{
A function to compare two COMPADRE/COMADRE database versions
}
\details{
\code{compareDBs} is preferred, but \code{dbCompare} is provided 
for legacy purposes.
}
\examples{
Compadre1 <- subset(Compadre, Continent == "Asia")
Compadre2 <- subset(Compadre, Continent == "Africa")

compareDBs(Compadre1, Compadre2)

}
\author{
Owen R. Jones <jones@biology.sdu.dk>
}
\keyword{attribute}
\keyword{manip}
