% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDBs.R
\name{compareDBs}
\alias{compareDBs}
\alias{dbCompare}
\title{Compare two versions or subsets of a COM(P)ADRE database}
\usage{
compareDBs(db1, db2, verbose = FALSE)

dbCompare(db1, db2, verbose = FALSE)
}
\arguments{
\item{db1, db2}{CompadreDB objects to compare}

\item{verbose}{A logical argument indicating whether or not to return lots
of detail}
}
\value{
NULL (output is printed rather than returned)
}
\description{
Prints a summary of the differences between two CompadreDB objects, including
the number of species, studies, and matrices in each. If argument
\code{verbose == TRUE}, additionally prints a list of the species and studies
that are present in one database but not the other.
}
\details{
\code{compareDBs} is preferred, but \code{dbCompare} is provided 
for legacy purposes.
}
\examples{
Compadre1 <- subset(Compadre, Continent == "Asia")
Compadre2 <- subset(Compadre, Continent == "Africa")

compareDBs(Compadre1, Compadre2)

}
\author{
Owen R. Jones <jones@biology.sdu.dk>
}
