% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDB.R
\name{convertLegacyDB}
\alias{convertLegacyDB}
\title{utility to convert older versions of database to
modern class}
\usage{
convertLegacyDB(db)
}
\arguments{
\item{db}{The \code{Com(p)adre} data base to convert}
}
\value{
A \code{CompadreData} object
}
\description{
utility to convert older versions of database to
modern class
}
