% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkspecies.R
\name{findSpecies}
\alias{findSpecies}
\title{Utility function for checkSpecies}
\usage{
findSpecies(x, db)
}
\arguments{
\item{x}{A character vector of species names.}

\item{db}{A COM(P)ADRE database object. Database versions <=4.0.1 will be coerced
from class 'list'.}
}
\value{
A logical indicating whether the species name is in the 
COM(P)ADRE object.
}
\description{
Utility function for checkSpecies
}
