% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanMatF.R
\name{getMeanMatF}
\alias{getMeanMatF}
\title{Calculates a population-specific mean fecundity matrix for each set of
matrices in a COM(P)ADRE database}
\usage{
getMeanMatF(db)
}
\arguments{
\item{db}{A CompadreDB object}
}
\value{
Returns a list of matrices, reprenting the mean fecundity matrix
  associated with each row of the database.
}
\description{
Takes a CompadreDB object and calculates a grand mean fecundity matrix for
each unique population (a mean of all population-specific fecundity matrices,
including fecundity matrices for which \code{MatrixComposite == 'Mean'}).
Here, a unique study population is defined as a unique combination of the
metadata columns 'Authors', 'YearPublication', 'DOI.ISBN', 'SpeciesAuthor',
'MatrixPopulation', and 'MatrixDimension'. The main purpose of this function
is to identify stage classes that are \emph{potentially} reproductive (i.e.
the absense of fecundity in a given stage class and year does not necessarily
indicate that the stage in question is non-reproductive).
}
\examples{
# print matF associated with row 16 of database
Compadre$mat[[16]]

# create list of meanMatFs
meanF <- getMeanMatF(Compadre)

# print meanMatF associated with row 16 of database
meanF[[16]]

}
\author{
Danny Buss <dlb50@cam.ac.uk>

Julia Jones <juliajones@biology.sdu.dk>

Iain Stott <stott@biolgy.ox.ac.uk>

Patrick Barks <patrick.barks@gmail.com>
}
