% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDBs.R
\name{mergeDBs}
\alias{mergeDBs}
\title{Merge two COM(P)ADRE databases together}
\usage{
mergeDBs(db1, db2)
}
\arguments{
\item{db1, db2}{CompadreDB objects}
}
\value{
A CompadreDB object containing both databases
}
\description{
Merges two CompadreDB objects via a row-bind of the data slots.
}
\examples{
Compadre1 <- subset(Compadre, Continent == "Asia")
Compadre2 <- subset(Compadre, Continent == "Africa")

mergeDBs(Compadre1, Compadre2)

}
\author{
Sam Levin
}
