% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDBs.R
\name{mergeDBs}
\alias{mergeDBs}
\title{Merge two Compadre/Comadre databases together}
\usage{
mergeDBs(db1, db2)
}
\arguments{
\item{db1}{A \code{CompadreDB} object}

\item{db2}{A \code{CompadreDB} object}
}
\value{
A \code{CompadreDB} object containing both databases.
}
\description{
Merges two CompadreDB objects together.
}
\examples{
Compadre1 <- subset(Compadre, Continent == "Asia")
Compadre2 <- subset(Compadre, Continent == "Africa")

mergeDBs(Compadre1, Compadre2)

}
\author{
Sam Levin
}
