% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDBs.R
\name{mergeDBs}
\alias{mergeDBs}
\title{Merge two Compadre/Comadre databases together}
\usage{
mergeDBs(db1, db2)
}
\arguments{
\item{db1}{A \code{CompadreDB} object}

\item{db2}{A \code{CompadreDB} object}
}
\value{
A \code{CompadreDB} object containing both databases.
}
\description{
Merges two CompadreDB objects together.
}
\examples{
\dontrun{
CompadreData(Compadre)
CompadreData(Comadre)

BigDB <- mergeDBs(Compadre, Comadre)
}

}
\author{
Sam Levin
}
